+++ src/utils/sdk.js
@@ -1,6 +1,6 @@
 /* eslint-disable eslint-comments/disable-enable-pair, no-else-return */
 import arrayFrom from 'core-js-pure/stable/array/from';
-import { getStorage as getBelterStorage } from '@krakenjs/belter/src';
+import { getStorage as getBelterStorage, uniqueID as createBelterID } from '@krakenjs/belter/src';
 import { SDK_QUERY_KEYS, SDK_SETTINGS } from '@paypal/sdk-constants/src';
 import {
     getClientID,
@@ -14,6 +14,8 @@ import {
     getCSPNonce,
     getNamespace as getSDKNamespace,
     getDefaultNamespace as getDefaultSDKNamespace,
+    getGlobalSessionID as getSDKGlobalSessionID,
+    setGlobalSessionID as setSDKGlobalSessionID,
     getSessionID as getSDKSessionID,
     getStorageID as getSDKStorageID,
     getStorageState as getSDKStorageState,
@@ -159,6 +161,23 @@ export function isZoidComponent() {
 
 export function getStorage() {
     return getBelterStorage({ name: getNamespace() });
+}
+
+// Uses SDK methods to get and set a global session ID
+// value will be passed in message_render events
+// and used to correlate with button events
+export function getOrCreateGlobalSessionID() {
+    if (__MESSAGES__.__TARGET__ === 'SDK') {
+        let globalSessionID = getSDKGlobalSessionID();
+        if (!globalSessionID) {
+            globalSessionID = createBelterID();
+            setSDKGlobalSessionID(globalSessionID);
+        }
+
+        return globalSessionID;
+    } else {
+        return getStorage().getSessionID();
+    }
 }
 
 // Use SDK methods when available, otherwise manually fetch storage via belter