+++ tests/functional/v2/config/index.js
@@ -6,3 +6,4 @@ export * as ES from './ES';
 export * as IT from './IT';
 export * as FR from './FR';
 export * as DE from './DE';
+export * as CA from './CA';